:zap: Filter backoff train data

+++ src/textchain/TextChain.hpp
@@ -20,11 +20,11 @@ public:
 
   [[nodiscard]] graph::Node *nextNode(const std::string &context) const;
 
-  [[nodiscard]] std::vector<graph::Node *> generateNodes(const std::string &context, uint32_t limit = UINT32_MAX) const;
+  [[nodiscard]] std::vector<graph::Node *> generateNodes(const std::string &context, uint32_t limit = INT8_MAX) const;
 
   [[nodiscard]] std::string nextToken(const std::string &context) const;
 
-  [[nodiscard]] std::string generateTokens(const std::string &context, uint32_t limit = UINT32_MAX,
+  [[nodiscard]] std::string generateTokens(const std::string &context, uint32_t limit = INT8_MAX,
                                            const std::string &breakValue = "") const;
 
 private:
